requirements.txt
# Core dependencies
openai>=1.12.0
chromadb>=0.4.22
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# MCP SDK
mcp>=0.9.0

# API framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0

# UI framework
streamlit>=1.31.0

# Utilities
tiktoken>=0.5.2
numpy>=1.26.0
aiofiles>=23.2.1

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-mock>=3.12.0

# Optional: Web search providers
tavily-python>=0.3.0

# Optional: Database support
sqlalchemy>=2.0.25

# Optional: Cloud storage
boto3>=1.34.0  # AWS S3
google-cloud-storage>=2.14.0  # GCS

# Optional: Snowflake
snowflake-connector-python>=3.7.0
pandas>=2.0.0  # For Snowflake data operations

